Bordure sauvegarde désormais sa largeur en valeur constante de pixels.

+++ Gaudrophone/src/gaudrophone/Domaine/Instrument/Bordure.java
@@ -5,7 +5,7 @@ import java.awt.Color;
 public class Bordure {
     int cote;
     boolean visible;
-    float largeur;
+    int largeur;
     Color couleur;
     
     public Bordure(){}
@@ -35,7 +35,7 @@ public class Bordure {
         return largeur;
     }
     
-    public void setLargeur(float largeur)
+    public void setLargeur(int largeur)
     {
         this.largeur = largeur;
     }
